story(ccls-1970): add applications query endpoint

+++ caab-service/src/main/java/uk/gov/laa/ccms/caab/api/entity/Application.java
@@ -60,7 +60,7 @@ public class Application implements Serializable {
 
   //Office
   @Column(name = "OFFICE_ID")
-  private Long officeId;
+  private Integer officeId;
   @Column(name = "OFFICE_DISPLAY_VALUE", length = 360)
   private String officeDisplayValue;
